feat: persist policies and ACL in SQLite instead of YAML config
Move dynamic state (policy rules, client ACL) from the YAML config file
to SQLite on the persistent volume. This fixes the core issue where
redeploying the Docker image would wipe all runtime policy changes.

- Add policy_rules and dynamic_config tables (migrations 12-13)
- Policy CRUD API now reads/writes SQLite directly
- First-boot seeds DB from YAML config for backward compatibility
- Policy IDs are now stable auto-increment integers (not array indices)
- Config watcher no longer hot-reloads policies (managed via API/DB)
- ACL API persists to SQLite with hot-reload to DNS server
- Client ACL settings page added to dashboard UI
- Entrypoint reverted to simple (no volume-copy hack needed)
- All 16 Go packages pass, 32 Playwright e2e tests pass

+++ pkg/api/handlers_policy_test.go
@@ -531,7 +531,7 @@ func TestHandleGetPolicies_MultiplePolicies(t *testing.T) {
 
 	// Verify all policies are present with correct IDs
 	for i, rule := range rules {
-		if result.Policies[i].ID != i {
+		if result.Policies[i].ID != int64(i) {
 			t.Errorf("Expected policy %d to have ID %d, got %d", i, i, result.Policies[i].ID)
 		}
